requirements.txt
-i https://pypi.org/simple
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; sys_platform == 'win32'
coverage[toml]==7.3.2; python_version >= '3.8'
cryptography==41.0.7; python_version >= '3.7'
greenlet==3.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
iniconfig==2.0.0; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
passlib==1.7.4
pluggy==1.3.0; python_version >= '3.8'
pycparser==2.21
pygments==2.17.2; python_version >= '3.7'
pyjwt[crypto]==2.8.0; python_version >= '3.7'
pymysql[rsa]==1.1.0; python_version >= '3.7'
pytest==7.4.3; python_version >= '3.7'
pytest-cov==4.1.0; python_version >= '3.7'
pytest-mock==3.12.0; python_version >= '3.8'
rich==13.7.0; python_full_version >= '3.7.0'
sentry-sdk==1.38.0
setuptools==69.0.2; python_version >= '3.8'
sqlalchemy==2.0.23; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.6'
